split paths package files
