+++ packages/paths/src/searchUp.ts
@@ -1,0 +1,25 @@
+import path from "path";
+import fs from "fs";
+
+/**
+ * Starting from `cwd`, searches up the directory hierarchy for `filePath`.
+ * If multiple strings are given, searches each directory level for any of them.
+ * @returns Full path to the item found, or undefined if not found.
+ */
+export function searchUp(filePath: string | string[], cwd: string) {
+  const paths = typeof filePath === "string" ? [filePath] : filePath;
+  const root = path.parse(cwd).root;
+
+  let foundPath: string | undefined;
+
+  while (!foundPath && cwd !== root) {
+    foundPath = paths.find((p) => fs.existsSync(path.join(cwd, p)));
+    if (foundPath) {
+      break;
+    }
+
+    cwd = path.dirname(cwd);
+  }
+
+  return foundPath ? path.join(cwd, foundPath) : undefined;
+}
